feat: refactor assignment workflow by consolidating date utilities and updating assignment components
