+++ web/src/components/features/operations/assignments/assignments-dashboard.tsx
@@ -95,8 +95,8 @@ export function AssignmentsDashboard({ data }: AssignmentsDashboardProps) {
     owner: "-",
     group: asset.pillar,
     assignedTo: asset.assignedTo ?? "-",
-    dateCreated: asset.returnedDate ? asset.returnedDate.toLocaleDateString("en-GB") : "-",
-    updatedAt: asset.returnedDate ? asset.returnedDate.toLocaleDateString("en-GB") : "-",
+    dateCreated: "-",
+    updatedAt: "-",
     warranty: "-",
     note: asset.location ?? "-",
     assetTag: asset.assetTag,
